fixes config due to change in carbon usage

+++ config/rememberable.php
@@ -1,6 +1,6 @@
 <?php
 
 return [
-    'cacheLifetime' => env('CACHE_LIFETIME', 3600),
+    'cacheLifetime' => (int) env('CACHE_LIFETIME', 3600),
     'keys' => ['id'],
 ];
